image_app/pyproject.toml
[tool.poetry]
name = "image-app"
version = "0.1.0"
description = ""
authors = ["Daniil Maximov <mcsdanf98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
pydantic-settings = "^2.6.1"
black = "^24.10.0"
pytest-asyncio = "0.23.6"
pytest = "8.1.1"
httpx = "^0.27.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
python-multipart = "^0.0.17"
aiofiles = "^24.1.0"
aiokafka = "^0.12.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
pydantic = {extras = ["email"], version = "^2.9.2"}

[tool.pytest.ini_options]
pythonpath = "app"
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py", "*_tests.py"]
asyncio_default_fixture_loop_scope = "function"
addopts = "-ra -q"
testpaths = [
    "tests",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
